chore(services): Break circular dependency between FlowsStore and StepsService
This situation happens when useFlowsStore imports FlowsService and StepsService
from @kaoto/services, and at the same time StepsService imports useFlowsStore.
