+++ src/services/stepsService.test.ts
@@ -17,7 +17,6 @@ import deployment from '../store/data/deployment';
 import nestedBranch from '../store/data/kamelet.nested-branch.steps';
 import steps from '../store/data/steps';
 import YAML from '../store/data/yaml';
-import { integrationSteps } from '../stubs';
 import { StepsService } from './stepsService';
 import {
   fetchDeployment,
@@ -198,11 +197,6 @@ describe('stepsService', () => {
 
     useFlowsStore.getState().deleteStep('Camel Route-1', 'Camel Route-1_pineapple-0');
     expect(useFlowsStore.getState().flows[0].steps).toHaveLength(0);
-  });
-
-  it('extractNestedSteps(): should create an array of properties for all nested steps', () => {
-    const nested = nestedBranch.slice();
-    expect(StepsService.extractNestedSteps(nested)).toHaveLength(6);
   });
 
   it("findStepIdxWithUUID(): should find a step's index, given a particular UUID", () => {
@@ -512,197 +506,6 @@ describe('stepsService', () => {
   it('isStartStep(): should determine if the provided step is a START step', () => {
     expect(StepsService.isStartStep(branchSteps[0])).toBe(true);
     expect(StepsService.isStartStep(branchSteps[1])).toBe(false);
-  });
-
-  describe('regenerateUuids(): should (re)generate UUIDs for all steps', () => {
-    it('should generate UUIDs for the main steps array', () => {
-      const result = StepsService.regenerateUuids('Camel Route-1', integrationSteps);
-
-      expect(result).toMatchObject<Partial<IStepProps>[]>([
-        { UUID: 'Camel Route-1_timer-0' },
-        { UUID: 'Camel Route-1_choice-1' },
-      ]);
-    });
-
-    it('should generate UUIDs for branches', () => {
-      const result = StepsService.regenerateUuids('Camel Route-1', integrationSteps);
-
-      expect(result).toMatchObject<Partial<IStepProps>[]>([
-        { UUID: 'Camel Route-1_timer-0', branches: [] },
-        {
-          UUID: 'Camel Route-1_choice-1',
-          branches: [
-            { branchUuid: 'Camel Route-1_choice-1_branch-0', steps: expect.any(Array), identifier: 'true path' },
-            { branchUuid: 'Camel Route-1_choice-1_branch-1', steps: expect.any(Array), identifier: 'otherwise' },
-          ],
-        },
-      ]);
-    });
-
-    it('should generate UUIDs for nested steps', () => {
-      const result = StepsService.regenerateUuids('Camel Route-1', integrationSteps);
-
-      expect(result).toMatchObject<Partial<IStepProps>[]>([
-        { UUID: 'Camel Route-1_timer-0', branches: [] },
-        {
-          UUID: 'Camel Route-1_choice-1',
-          branches: [
-            {
-              branchUuid: 'Camel Route-1_choice-1_branch-0',
-              steps: [
-                {
-                  UUID: 'Camel Route-1_choice-1_branch-0_log-0',
-                  integrationId: 'Camel Route-1',
-                  minBranches: 0,
-                  maxBranches: 0,
-                  name: 'log',
-                  type: 'MIDDLE',
-                },
-              ],
-              identifier: 'true path',
-            },
-            {
-              branchUuid: 'Camel Route-1_choice-1_branch-1',
-              steps: [
-                {
-                  UUID: 'Camel Route-1_choice-1_branch-1_log-0',
-                  integrationId: 'Camel Route-1',
-                  minBranches: 0,
-                  maxBranches: 0,
-                  name: 'log',
-                  type: 'MIDDLE',
-                },
-              ],
-              identifier: 'otherwise',
-            },
-          ],
-        },
-      ]);
-    });
-
-    it('should generate UUIDs for nested steps at two levels deep', () => {
-      const nestedIntegrationSteps: IStepProps[] = JSON.parse(JSON.stringify(integrationSteps));
-      nestedIntegrationSteps[1].branches![0].steps = integrationSteps;
-
-      const result = StepsService.regenerateUuids('Camel Route-1', nestedIntegrationSteps);
-
-      expect(result).toMatchObject<Partial<IStepProps>[]>([
-        { UUID: 'Camel Route-1_timer-0', branches: [] },
-        {
-          UUID: 'Camel Route-1_choice-1',
-          branches: [
-            {
-              branchUuid: 'Camel Route-1_choice-1_branch-0',
-              steps: [
-                {
-                  UUID: 'Camel Route-1_choice-1_branch-0_timer-0',
-                  integrationId: 'Camel Route-1',
-                  minBranches: 0,
-                  maxBranches: 0,
-                  name: 'timer',
-                  type: 'START',
-                  branches: [],
-                },
-                {
-                  UUID: 'Camel Route-1_choice-1_branch-0_choice-1',
-                  integrationId: 'Camel Route-1',
-                  minBranches: 1,
-                  maxBranches: -1,
-                  name: 'choice',
-                  type: 'MIDDLE',
-                  branches: [
-                    {
-                      branchUuid: 'Camel Route-1_choice-1_branch-0_choice-1_branch-0',
-                      steps: [
-                        {
-                          UUID: 'Camel Route-1_choice-1_branch-0_choice-1_branch-0_log-0',
-                          integrationId: 'Camel Route-1',
-                          minBranches: 0,
-                          maxBranches: 0,
-                          name: 'log',
-                          type: 'MIDDLE',
-                        },
-                      ],
-                      identifier: 'true path',
-                    },
-                    {
-                      branchUuid: 'Camel Route-1_choice-1_branch-0_choice-1_branch-1',
-                      steps: [
-                        {
-                          UUID: 'Camel Route-1_choice-1_branch-0_choice-1_branch-1_log-0',
-                          integrationId: 'Camel Route-1',
-                          minBranches: 0,
-                          maxBranches: 0,
-                          name: 'log',
-                          type: 'MIDDLE',
-                        },
-                      ],
-                      identifier: 'otherwise',
-                    },
-                  ],
-                },
-              ],
-              identifier: 'true path',
-            },
-            {
-              branchUuid: 'Camel Route-1_choice-1_branch-1',
-              steps: [
-                {
-                  UUID: 'Camel Route-1_choice-1_branch-1_log-0',
-                  integrationId: 'Camel Route-1',
-                  minBranches: 0,
-                  maxBranches: 0,
-                  name: 'log',
-                  type: 'MIDDLE',
-                },
-              ],
-              identifier: 'otherwise',
-            },
-          ],
-        },
-      ]);
-    });
-
-    it('should regenerate UUIDs when a branch is removed', () => {
-      const localIntegrationSteps = integrationSteps.slice(1);
-      const result = StepsService.regenerateUuids('Camel Route-1', localIntegrationSteps);
-
-      expect(result).toMatchObject<Partial<IStepProps>[]>([
-        {
-          UUID: 'Camel Route-1_choice-0',
-          branches: [
-            {
-              branchUuid: 'Camel Route-1_choice-0_branch-0',
-              steps: [
-                {
-                  UUID: 'Camel Route-1_choice-0_branch-0_log-0',
-                  integrationId: 'Camel Route-1',
-                  minBranches: 0,
-                  maxBranches: 0,
-                  name: 'log',
-                  type: 'MIDDLE',
-                },
-              ],
-              identifier: 'true path',
-            },
-            {
-              branchUuid: 'Camel Route-1_choice-0_branch-1',
-              steps: [
-                {
-                  UUID: 'Camel Route-1_choice-0_branch-1_log-0',
-                  integrationId: 'Camel Route-1',
-                  minBranches: 0,
-                  maxBranches: 0,
-                  name: 'log',
-                  type: 'MIDDLE',
-                },
-              ],
-              identifier: 'otherwise',
-            },
-          ],
-        },
-      ]);
-    });
   });
 
   it('replacePlaceholderStep(): should replace a placeholder with a step and update the store', async () => {
