+++ src/services/stepsService.ts
@@ -24,17 +24,14 @@ import {
   IVizStepNodeData,
 } from '@kaoto/types';
 import { findPath, getDeepValue } from '@kaoto/utils';
-import cloneDeep from 'lodash.clonedeep';
 
 /**
  * A collection of business logic to handle logical model objects of the integration,
  * which is represented by a collection of "Step".
  * This class focuses on handling logical model objects. For handling visualization,
- * see {@link VisualizationService}.
  * Note: Methods are declared in alphabetical order.
  * @see IStepProps
  * @see IStepPropsBranch
- * @see VisualizationService
  */
 export class StepsService {
 
@@ -172,50 +169,6 @@ export class StepsService {
       useVisualizationStore.getState().deleteNode(stepsIndex);
       useFlowsStore.getState().deleteStep(integrationId, UUID);
     }
-  }
-
-  /**
-   * Given an array of Steps, return an array containing *only*
-   * the steps which are nested
-   * @param steps
-   */
-  static extractNestedSteps(steps: IStepProps[]) {
-    let tempSteps = steps.slice();
-    let nestedSteps: INestedStep[] = [];
-
-    const loopOverSteps = (
-      steps: IStepProps[],
-      parentStepUuid?: string,
-      branchUuid?: string,
-      branchIdx?: number
-    ) => {
-      steps.forEach((step) => {
-        if (parentStepUuid) {
-          // this is a nested step
-          nestedSteps.push({
-            branchIndex: branchIdx ?? undefined,
-            branchUuid,
-            stepUuid: step.UUID,
-            parentStepUuid,
-            pathToBranch: branchUuid ? findPath(tempSteps, branchUuid, 'branchUuid') : undefined,
-            pathToParentStep: findPath(tempSteps, parentStepUuid, 'UUID'),
-            pathToStep: findPath(tempSteps, step.UUID, 'UUID'),
-          } as INestedStep);
-        }
-
-        if (this.supportsBranching(step) && this.containsBranches(step)) {
-          step.branches!.forEach((branch, branchIdx) => {
-            // it contains nested steps; we will need to store the branch info
-            // and the path to it, for each of those steps
-            return loopOverSteps(branch.steps, step.UUID, branch.branchUuid, branchIdx);
-          });
-        }
-      });
-    };
-
-    loopOverSteps(tempSteps);
-
-    return nestedSteps;
   }
 
   /**
@@ -465,30 +418,6 @@ export class StepsService {
 
   static isStartStep(step: IStepProps): boolean {
     return step.type === 'START';
-  }
-
-  /**
-   * Regenerate a UUID for a list of Steps
-   * Every time there is a change to steps or their positioning in the Steps array,
-   * their UUIDs need to be regenerated
-   * @param steps
-   * @param prefix
-   */
-  static regenerateUuids(integrationId: string, steps: IStepProps[], prefix?: string): IStepProps[] {
-    let newSteps = cloneDeep(steps);
-    const integrationPrefix = prefix ?? `${integrationId}_`;
-
-    newSteps.forEach((step, stepIndex) => {
-      step.UUID = `${integrationPrefix}${step.name}-${stepIndex}`;
-      step.integrationId = integrationId;
-
-      step.branches?.forEach((branch, branchIndex) => {
-        branch.branchUuid = `${step.UUID}_branch-${branchIndex}`;
-        branch.steps = StepsService.regenerateUuids(integrationId, branch.steps, `${branch.branchUuid}_`)
-      });
-    });
-
-    return newSteps;
   }
 
   /**
